chatbot_api/pyproject.toml
[tool.poetry]
name = "API Chatbot  Project CS"
version = "0.2.0"
description = "API Chatbot LLM (LLM Large language model)  with  RAG (Retrieval-Augmented Generation) for RMUTL NAN"
authors = ["Jeerasak Ananta <jeerasakananta@gmail.com>"]
maintainers = ["Jeerasak Ananta <jeerasakananta@gmail.com>"]
license = "MIT"
readme = "README.md"

#  dependencies
[tool.poetry.dependencies]
python = "^3.12"
langchain-community = "0.2.16"
langchain-openai = "0.1.23"
python-dotenv = "1.0.1"
pymupdf = "1.25.1"
qdrant-client = "1.12.1"
langchain-qdrant = "0.1.3"
jinja2 = "3.1.4"
python-multipart = "0.0.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
docling = "^2.29.0"


#  dev dependencies 
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
